+++ ding/model/template/__init__.py
@@ -31,4 +31,4 @@ from .edac import EDAC
 from .qgpo import QGPO
 from .ebm import EBM, AutoregressiveEBM
 from .havac import HAVAC
-from .policy_stem import PolicyStem
+
