+++ dizoo/box2d/lunarlander/entry/lunarlander_hpt_example.py
@@ -1,5 +1,7 @@
 
 import gym
+import torch
+import torch.nn as nn
 from ditk import logging
 from ding.data.model_loader import FileModelLoader
 from ding.data.storage_loader import FileStorageLoader
@@ -16,8 +18,7 @@ from ding.framework.middleware import OffPolicyLearner, StepCollector, interacti
     nstep_reward_enhancer
 from ding.utils import set_pkg_seed
 from dizoo.box2d.lunarlander.config.lunarlander_hpt_config import main_config, create_config
-import torch
-import torch.nn as nn
+
 
 
 
@@ -38,24 +39,13 @@ def main():
 
         set_pkg_seed(cfg.seed, use_cuda=cfg.policy.cuda)
 
-        # 迁移模型到 GPU
+        # Migrating models to the GPU
         device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-        # model = DQN(**cfg.policy.model).to(device)
         model = HPT(cfg.policy.model.obs_shape,cfg.policy.model.action_shape).to(device)
-
-
-
-         # 检查模型是否在 GPU
-        for param in model.parameters():
-            print("模型参数所在设备：", param.device)
-            break
         buffer_ = DequeBuffer(size=cfg.policy.other.replay_buffer.replay_buffer_size)
 
-        # 将模型传入 Policy
+        # Pass the model into Policy
         policy = DQNPolicy(cfg.policy, model=model) 
-        print("日志保存路径：", cfg.exp_name)
-
 
         # Consider the case with multiple processes
         if task.router.is_active:
@@ -74,7 +64,6 @@ def main():
         
 
         # Here is the part of single process pipeline.
-        # evaluator_env.enable_save_replay(replay_path='./video')
         task.use(interaction_evaluator(cfg, policy.eval_mode, evaluator_env))
         task.use(eps_greedy_handler(cfg))
         task.use(StepCollector(cfg, policy.collect_mode, collector_env))
